Version 1.03

+++ includes/deletenotice.php
@@ -1,12 +1,13 @@
 <?php
-
 session_start();
 include("config.php");
 
+if ( $_SESSION["level"] == '2')
+{
 $ID = $_GET['ID'];
 
-//if ( $_SESSION['loginok'] == 'ok' && $_SESSION['level'] == '2' ) {
-	mysql_query("DELETE FROM notices WHERE ID = $ID ",$con);
-	header("Location: http://www.iiitcslcentral.co.cc/");
-//}
+mysql_query("DELETE FROM notices WHERE ID = $ID ",$con);
+header("Location: http://www.iiitcslcentral.co.cc/");
+
+}
 ?>
